requirements.txt
# Core Framework
flask==3.1.1
python-dotenv==1.1.0
gunicorn==23.0.0

# AI / ML
tensorflow>=2.15.0
torch>=2.0.0
torchvision>=0.15.0
google-generativeai>=0.8.0

# Image Processing
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Report Generation
fpdf2>=2.8.0
markdown>=3.5.0

# Utilities
requests>=2.31.0
